Moved dialogs to their own seperate files and added more documentation
