+++ generators/app/index.js
@@ -23,8 +23,8 @@ module.exports = class extends Generator {
     const directoryName = _.kebabCase(this.props.botName);
     const extension = this.props.language === 'JavaScript' ? 'js' : 'ts';
     const launchSteps = extension === 'js' ? `node app.js` : `tsc\nnode app.js`;
-    const defaultDialog = this.props.dialog.split(' ')[0].toLowerCase() + `Dialog`;
-    const luisRegistration = (defaultDialog === 'luisDialog') ? '\nbot.recognizer(new builder.LuisRecognizer(process.env.LUIS_MODEL_URL));\n' : '\n';
+    const defaultDialog = this.props.dialog.split(' ')[0].toLowerCase();
+    const luisRegistration = (defaultDialog === 'luis') ? '\nbot.recognizer(new builder.LuisRecognizer(process.env.LUIS_MODEL_URL));\n' : '\n';
 
     if (path.basename(this.destinationPath()) !== directoryName) {
       this.log(`Your bot should be in a directory named ${directoryName}\nI'll automatically create this folder.`);
@@ -41,7 +41,7 @@ module.exports = class extends Generator {
     this.fs.copyTpl(this.templatePath(`bot.${extension}`), this.destinationPath(`bot.${extension}`), {
       defaultDialog, luisRegistration
     });
-    this.fs.copyTpl(this.templatePath(`dialogs.${extension}`), this.destinationPath(`dialogs.${extension}`), {
+    this.fs.copyTpl(this.templatePath(`dialogs-${extension}`), this.destinationPath(`dialogs`), {
       botName: this.props.botName, botDescription: this.props.description
     });
