+++ generators/app/templates/bot.js
@@ -1,13 +1,13 @@
 "use strict";
 
 const builder = require("botbuilder");
-const dialogs = require("./dialogs");
+const dialog = require("./dialogs/<%= defaultDialog %>");
 const bot = new builder.UniversalBot(
     new builder.ChatConnector({
         appId: process.env.MICROSOFT_APP_ID,
         appPassword: process.env.MICROSOFT_APP_PASSWORD
     }), 
-    dialogs.<%= defaultDialog %>.dialog
+    dialog.waterfall
 );
 <%= luisRegistration %>
 module.exports = bot;
