+++ generators/app/templates/dialogs-js/qna.js
@@ -1,23 +1,29 @@
-import * as restify from 'restify';
-import * as builder from 'botbuilder';
-import { IDialog } from './idialog';
+/* ----------------------------------------------------------------------------------
+*   QnA Dialog
+*   Sample dialog for use with QnA Maker 
+*   You can find out more information about QnA Maker at https://qnamaker.ai
+*       or at https://channel9.msdn.com/Events/Build/2017/p4074
+*
+*   To use:
+*   1. Create a knowledge base at https://qnamaker.ai
+*   2. Set the KBID and SUBSCRIPTION_KEY environmental variables
+*       This can be done by updating .env, or setting the variables manually
+*   3. No additional code updates are required
+---------------------------------------------------------------------------------- */
 
-interface IQnAAnswer {
-    answer: string;
-    score: number;
-}
+const builder = require('botbuilder');
+const restify = require('restify');
 
-export const qnaDialog: IDialog = {
+module.exports = {
     id: 'qna',
     name: 'qna',
     waterfall: [
-        (session: builder.Session, args: any, next: Function) => {
+        (session, args, next) => {
             const question = session.message.text;
             if (!question) {
                 session.endConversation(`Sorry, I can't answer that question.`);
                 return;
             }
-
             session.sendTyping();
             const bodyText = JSON.stringify({ question: question });
             const url = `https://westus.api.cognitive.microsoft.com`;
@@ -28,22 +34,22 @@ export const qnaDialog: IDialog = {
                     'Ocp-Apim-Subscription-Key': process.env.SUBSCRIPTION_KEY
                 }
             });
-
-            jsonClient.post(path, { question }, (err, req, res, obj: IQnAAnswer) => {
+            jsonClient.post(path, { question }, (err, req, res, obj) => {
                 if (err) {
                     session.endConversation(`Sorry, something went wrong when connecting to the server`);
                     return;
                 }
-
                 if (obj.score > 50) {
                     session.endConversation(obj.answer);
-                } else if (obj.score > 0) {
+                }
+                else if (obj.score > 0) {
                     session.send(`I'm not sure if this is right...`);
                     session.endConversation(obj.answer);
-                } else {
+                }
+                else {
                     session.endConversation(`I don't have that answer.`);
                 }
             });
         }
     ]
-}
+};
