+++ generators/app/templates/dialogs-ts/idialog.ts
