+++ generators/app/templates/dialogs-ts/qna.ts
@@ -1,55 +1,29 @@
+/* ----------------------------------------------------------------------------------
+*   QnA Dialog
+*   Sample dialog for use with QnA Maker 
+*   You can find out more information about QnA Maker at https://qnamaker.ai
+*       or at https://channel9.msdn.com/Events/Build/2017/p4074
+*
+*   To use:
+*   1. Create a knowledge base at https://qnamaker.ai
+*   2. Set the KBID and SUBSCRIPTION_KEY environmental variables
+*       This can be done by updating .env, or setting the variables manually
+*   3. No additional code updates are required
+---------------------------------------------------------------------------------- */
+
 import * as restify from 'restify';
 import * as builder from 'botbuilder';
-
-export interface IDialog {
-    id: String;
-    name: String;
-    dialog: builder.IDialogWaterfallStep[];
-}
-
-export const echoDialog: IDialog = {
-    id: 'echo',
-    name: 'echo',
-    dialog: [
-        (session, args, next) => {
-            const botName = '<%= botName %>';
-            const description = `<%= botDescription %>`;
-
-            session.send(`Hi there! I'm ${botName}`);
-            session.send(`In a nutshell, here's what I can do:\n\n${description}`);
-
-            builder.Prompts.text(session, `What's your name?`);
-        },
-        (session, results, next) => {
-            session.endConversation(`Welcome, ${results.response}`);
-        },
-    ]
-}
-
-export const luisDialog: IDialog = {
-    id: 'none',
-    name: 'none',
-    dialog: [
-        (session, args, next) => {
-            const entity = builder.EntityRecognizer.findEntity(args.entities, 'entity');
-            if(entity) next({ response: entity.entity });
-            else builder.Prompts.text(session, 'Please provide entityName');
-        },
-        (session, results, next) => {
-            session.endConversation(`You said ${results.response}`);
-        }
-    ]
-}
+import { IDialog } from './idialog';
 
 interface IQnAAnswer {
     answer: string;
     score: number;
 }
 
-export const qnaDialog: IDialog = {
+const dialog: IDialog = {
     id: 'qna',
     name: 'qna',
-    dialog: [
+    waterfall: [
         (session: builder.Session, args: any, next: Function) => {
             const question = session.message.text;
             if (!question) {
@@ -74,10 +48,12 @@ export const qnaDialog: IDialog = {
                     return;
                 }
 
+                // The code below sets a threshold at 50 when sending the reply back.
+                // You can update (or remove) that threshold as you see fit
                 if (obj.score > 50) {
                     session.endConversation(obj.answer);
                 } else if (obj.score > 0) {
-                    session.send(`I'm not sure if this is right...`);
+                    session.send(`I'm not sure if this is right, but here's what I know...`);
                     session.endConversation(obj.answer);
                 } else {
                     session.endConversation(`I don't have that answer.`);
@@ -86,3 +62,5 @@ export const qnaDialog: IDialog = {
         }
     ]
 }
+
+export default dialog;
